chore!: remove the `--testnet` argument
BREAKING CHANGE: the `ant`, `antctl` and `antnode` binaries will no longer support the `--testnet`
argument.

This argument will no longer be required because we can designate we are using a testnet by setting
`--network-id` to any value above 2, where 1 will be reserved for `mainnet` and 2 will be reserved
for `alphanet`.

Due to maintaining backwards compatibility with the node service data that is used by `antctl`, we
had to introduce backwards compatibility with the `InitialPeersConfig` struct, thus it now has two
versions. A new version of `NodeServiceData` was introduced for the removal of the argument, and we
will also add a new `--alpha` flag argument to this same version.