chore: standardise autonomi usage in connection messages
Our network is called 'Autonomi', not 'the Autonomi Network', and therefore only 'Autonomi' should
be capitalised to denote it as a proper noun.

The words 'alpha', 'local', and 'custom' are variants of Autonomi networks, so they should not be
capitalised.

When we're connecting to the production network, we can just say 'Connecting to Autonomi', without
the word 'network'.

Other usages of the word 'network' being capitalised were changed because 'network' is not a proper
noun in that context.