chore: utility target for providing binary hashes
The automatic upgrading process is going to require the sha256 hashes of the binaries be included in
the information for each release.

This new target in the Justfile can operate on the populated `artifacts` directory to get all the
hashes. It will output them in a markdown table which can be pasted into the release description.