chore!: use 'always' restart policy for `antctl add`
BREAKING CHANGE: new node deployments with `antctl` will apply a different restart policy.

With `antctl` we are going to configure services for node deployments to always restart. This is so
we can use the node deployments with the automatic upgrades feature and let the service manager
handle the restart of the `antnode` process. If `antnode` itself were to restart the process, the
service manager would consider it to be dead, which would break the status tracking for `antctl`.

We need to provide a `--stop-on-upgrade` argument for `antnode` such that it will stop the process
on the upgrade rather than restart it, and allow the service manager to perform the restart. This
will be coming in the next commit.

The `node-launchpad` has been configured to use 'never' for the restart policy because we ran into
problems with the UPnP feature if the launchpad continually restarted nodes. So automatic upgrades
just won't be supported when using the launchpad. We will need to provide another `antnode` argument
to disable automatic upgrades, which will come in another commit.