feat: provide `--stop-on-upgrade` argument for `antctl`
If this argument is used the node will be stopped after it is upgraded, rather than restarted.

This is to allow the service manager to restart the process if the node is being run in this
context, which it will be if it is deployed using `antctl`, which is updated in this commit to
always supply the argument as part of the service definition for a node.