feat: support restarting symlinked binaries
The code performing restarts of `antnode` was using the canonical path of the executable, which in
the case of symlinks, is the path of the source of the link rather than the link itself.

To support a symlinked binary setup in automatic upgrades, we now modify the restart mechanism to
use the invoked path, which is the location of the symlink. For non-symlinked binaries the mechanism
should work as normal.

I've also removed use of `retain_peer_id`, which appended `--root-dir` and `--port` arguments to the
end of the new `antnode` invocation. This is an option on the RPC command, which will soon be phased
out anyway. It was also causing the new node process to exit immediately with a failure, since the
order of arguments matters when `antnode` is being invoked: the EVM options must be at the end. We
can just launch the restarted node with the same arguments it was originally invoked with.