chore: periodically refresh node registry
Now that we have automatic upgrades, a background task will run on `node-launchpad` to refresh the
node registry every 3 minutes.

The refresh mechanism has already been updated to detect the new version of services, so the node
registry will get re-written with the new version and the UI will be updated.