chore: retain 'never' restart policy on windows
We changed this policy to 'on success' on Linux and macOS to facilitate automatic upgrades with node
deployments setup through `antctl`.

Windows doesn't have an 'on success' equivalent, so it will revert to 'always', which is not what we
want. We use conditional compilation to set the policy back to 'never' for Windows and retain the
original behaviour.